+++ mayan/apps/file_metadata/classes.py
@@ -335,10 +335,16 @@ class FileMetadataDriver(
 
             for key, value in file_metadata_dictionary.items():
                 internal_name = internal_name_dictionary_deduplicated[key]
+
+                # Drivers should not be returning `None` values.
+                # Added to workaround undocumented backward incompatible
+                # changes in Ollama.
+                value_clean = value or ''
+
                 coroutine.send(
                     {
                         'document_file_driver_entry': document_file_driver_entry,
-                        'internal_name': internal_name, 'key': key, 'value': value
+                        'internal_name': internal_name, 'key': key, 'value': value_clean
                     }
                 )